pyproject.toml
[tool.poetry]
name = "swarmalators"
version = "0.1.0"
description = ""
authors = ["Richard Beattie <richard.beattie91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.8.0.76"
pyserial = "^3.5"
numpy = "^1.25.2"
filterpy = "1.4.5"
scikit-image = "^0.21.0"
lapx = "^0.5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
